// fahrschule-autopilot-website/src/components/anmeldung/types.ts
export interface FormData {
  vorname: string;
  nachname: string;
  geburtsdatum: string;
  telefon: string;
  email: string;
  strasse: string;
  plz: string;
  ort: string;
  klasse: string;
  vorbesitz: string[];
  sehtest: string;
  ersteHilfe: string;
  theoriesprache: string;
  bevorzugteTage: string[];
  bevorzugteUhrzeit: string;
  wieErfahren: string;
  nachricht: string;
  dsgvo: boolean;
  kontaktEinwilligung: boolean;
}